feat!: Use either V2 or V3 aws sdk
