+++ src/adapters/helpers/requestError.ts
@@ -1,4 +1,3 @@
-import { InvocationRequest, InvocationResponse } from '@aws-sdk/client-lambda';
 import { HandlerRequest, AlphaOptions } from '../../types';
 import { AxiosError, AxiosResponse } from 'axios';
 
@@ -8,7 +7,10 @@ export const isAxiosError = (err: any | AxiosError): err is AxiosError =>
 export const isAlphaRequestError = (err: any | RequestError): err is RequestError =>
   (typeof err === 'object') && !!err.isAlphaRequestError;
 
-export class RequestError extends Error implements Omit<AxiosError, 'response' | 'toJSON' | 'isAxiosError'> {
+export class RequestError<
+  Request extends Record<string, any> = Record<string, any>,
+  Response extends Record<string, any> = Record<string, any>
+> extends Error implements Omit<AxiosError, 'response' | 'toJSON' | 'isAxiosError'> {
   public code?: string;
   public isLambdaInvokeTimeout?: boolean;
   public isAlphaRequestError = true;
@@ -16,8 +18,8 @@ export class RequestError extends Error implements Omit<AxiosError, 'response' |
   constructor (
     message: string,
     public config: AlphaOptions,
-    public request: InvocationRequest | HandlerRequest,
-    public response?: InvocationResponse | AxiosResponse,
+    public request: Request | HandlerRequest,
+    public response?: Response | AxiosResponse,
   ) {
     super(message);
   }
