+++ src/awsV3/credentialsV3.ts
@@ -1,0 +1,4 @@
+import { defaultProvider } from '@aws-sdk/credential-provider-node';
+import { Credentials, Provider } from '@aws-sdk/types';
+
+export const credentialsProvider: Provider<Credentials> = defaultProvider();
