+++ src/types.ts
@@ -1,5 +1,4 @@
 import type { AxiosPromise, AxiosRequestConfig, AxiosResponse } from 'axios';
-import type { Lambda } from '@aws-sdk/client-lambda';
 import type { Context, Handler } from 'aws-lambda';
 import { SignatureV4CryptoInit, SignatureV4Init } from '@aws-sdk/signature-v4';
 
@@ -13,24 +12,25 @@ export interface RetryOptions {
 type SignatureV4Constructor = SignatureV4Init & SignatureV4CryptoInit;
 type SignatureV4Optionals = 'credentials' | 'region' | 'sha256' | 'service';
 
-export interface AlphaResponse<ResponseData = any, ConfigData = any> extends AxiosResponse<ResponseData> {
-  config: AlphaOptions<ConfigData>;
-}
-
 export type SignAwsV4Config =
   & Omit<SignatureV4Constructor, SignatureV4Optionals>
   & Partial<Pick<SignatureV4Constructor, SignatureV4Optionals>>;
 
-export interface AlphaOptions<D = any> extends AxiosRequestConfig<D> {
+export interface AlphaResponse<ResponseData = any, ConfigData = any> extends AxiosResponse<ResponseData> {
+  config: AlphaOptions<ConfigData>;
+}
+
+export interface AlphaOptions<D = any, T extends any = any> extends AxiosRequestConfig<D> {
   retry?: RetryOptions | boolean;
   lambda?: Handler;
   context?: Context;
-  signAwsV4?: SignAwsV4Config;
   /**
    * (Optional) The AWS endpoint to use when invoking the target Lambda function.
    */
   lambdaEndpoint?: string;
-  Lambda?: typeof Lambda;
+  Lambda?: T;
+  signAwsV4?: SignAwsV4Config;
+  awsSdkVersion?: 2 | 3;
 }
 
 export type AlphaAdapter = (config: AlphaOptions) => AxiosPromise;
